+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 - New card "Blizzard of Spears" (thanks @dilawer007)
 - Fixed a loophole that allowed "Grinning Hunger" to avoid its mandatory effect by choosing an empty zone (thanks @DragosIonita23)
 - Updated multipart card selections to automatically resolve mandatory selections with only one possible outcome (thanks @DragosIonita23)
+- New card "Balza, Seeker of Hyperpearls" (thanks @DragosIonita23)
+- New card "Ryudmila, Channeler of Suns" (thanks @DragosIonita23)
+- New card "King Oquanos" (thanks @DragosIonita23)
+- New card "Gajirabute, Vile Centurion" (thanks @DragosIonita23)
+- New card "Gaulezal Dragon" (thanks @DragosIonita23)
+- New card "Carnival Totem" (thanks @DragosIonita23)
 - Brand new website
 - Added support for playing as a guest without registering an account
 - Added email verification for new users